BE:fmt: restructured and separated validation/domain schemas
