+++ backend/src/api/main.py
@@ -15,7 +15,7 @@ from fastapi import (
 from sqlalchemy import func, select
 from sqlalchemy.ext.asyncio import AsyncSession
 
-import src.models as d  # d - database
+import src.database as d  # d - database
 import src.schemas as s  # s - schema
 from config import Config, get_config
 from src.connection_manager import ConnectionManager
