+++ backend/src/schemas/__init__.py
@@ -1,3 +1,4 @@
 # Expose all schemas through this package under `s.*` namespace.
-from src.schemas.general import *  # noqa: F403
+from src.schemas.domain import *  # noqa: F403
+from src.schemas.validation import *  # noqa: F403
 from src.schemas.websockets import *  # noqa: F403
