+++ backend/word_chain_game.py
@@ -8,9 +8,9 @@ from fastapi.middleware.cors import CORSMiddleware
 
 from config import LOGGING_CONFIG
 from src.api import main, rooms
+from src.database import create_root_objects, recreate_database
 from src.error_handlers import request_validation_handler
 from src.helpers import tags_metadata
-from src.models import create_root_objects, recreate_database
 
 
 @asynccontextmanager
